requirements.txt
streamlit
pillow
anthropic
google-generativeai